owner: implement a compatibility layer between the old scheduler and the new one
