+++ cdc/owner/changefeed.go
@@ -17,6 +17,8 @@ import (
 	"context"
 	"sync"
 
+	schedulerv2 "github.com/pingcap/ticdc/cdc/scheduler"
+
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/log"
@@ -38,7 +40,7 @@ type changefeed struct {
 	id    model.ChangeFeedID
 	state *orchestrator.ChangefeedReactorState
 
-	scheduler        *scheduler
+	scheduler        scheduler
 	barriers         *barriers
 	feedStateManager *feedStateManager
 	gcManager        gc.Manager
@@ -76,7 +78,7 @@ type changefeed struct {
 func newChangefeed(id model.ChangeFeedID, gcManager gc.Manager) *changefeed {
 	c := &changefeed{
 		id:               id,
-		scheduler:        newScheduler(),
+		scheduler:        newSchedulerV1(),
 		barriers:         newBarriers(),
 		feedStateManager: new(feedStateManager),
 		gcManager:        gcManager,
@@ -178,14 +180,20 @@ func (c *changefeed) tick(ctx cdcContext.Context, state *orchestrator.Changefeed
 		// So we return here.
 		return nil
 	}
-	shouldUpdateState, err := c.scheduler.Tick(c.state, c.schema.AllPhysicalTables(), captures)
+	newCheckpointTs, newResolvedTs, err := c.scheduler.Tick(c.state, c.schema.AllPhysicalTables(), captures)
 	if err != nil {
 		return errors.Trace(err)
 	}
-	if shouldUpdateState {
+	if newCheckpointTs != schedulerv2.CheckpointCannotProceed {
 		pdTime, _ := ctx.GlobalVars().TimeAcquirer.CurrentTimeFromCached()
 		currentTs := oracle.GetPhysical(pdTime)
-		c.updateStatus(currentTs, barrierTs)
+		if newResolvedTs > barrierTs {
+			newResolvedTs = barrierTs
+		}
+		if newCheckpointTs > barrierTs {
+			newCheckpointTs = barrierTs
+		}
+		c.updateStatus(currentTs, newCheckpointTs, newResolvedTs)
 	}
 	return nil
 }
@@ -462,26 +470,7 @@ func (c *changefeed) asyncExecDDL(ctx cdcContext.Context, job *timodel.Job) (don
 	return done, nil
 }
 
-func (c *changefeed) updateStatus(currentTs int64, barrierTs model.Ts) {
-	resolvedTs := barrierTs
-	for _, position := range c.state.TaskPositions {
-		if resolvedTs > position.ResolvedTs {
-			resolvedTs = position.ResolvedTs
-		}
-	}
-	for _, taskStatus := range c.state.TaskStatuses {
-		for _, opt := range taskStatus.Operation {
-			if resolvedTs > opt.BoundaryTs {
-				resolvedTs = opt.BoundaryTs
-			}
-		}
-	}
-	checkpointTs := resolvedTs
-	for _, position := range c.state.TaskPositions {
-		if checkpointTs > position.CheckPointTs {
-			checkpointTs = position.CheckPointTs
-		}
-	}
+func (c *changefeed) updateStatus(currentTs int64, checkpointTs, resolvedTs model.Ts) {
 	c.state.PatchStatus(func(status *model.ChangeFeedStatus) (*model.ChangeFeedStatus, bool, error) {
 		changed := false
 		if status.ResolvedTs != resolvedTs {
